requirements.txt
Flask==1.1.1
google-api-core==1.14.3
google-auth==1.7.2
google-cloud-bigquery==1.22.0
google-cloud-core==1.1.0
google-resumable-media==0.5.0
googleapis-common-protos==1.6.0
html5lib==1.0.1
imagesize==1.1.0
importlib-metadata==0.17
ipykernel==5.1.1
ipython==7.6.1
ipython-genutils==0.2.0
isort==4.3.21
Jinja2==2.10.1
joblib==0.13.2
json5==0.8.4
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
jupyterlab==1.0.2
jupyterlab-server==1.0.0
matplotlib==3.1.2
numpy==1.16.4
numpydoc==0.9.1
pandas==0.24.2
py==1.8.0